backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
pandas
python-multipart
openpyxl
langchain[openai]
openai